DSDeepParser-v2-Claude/requirements.txt
aiohttp>=3.9,<4
pyyaml>=6.0
tenacity>=8.2
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
beautifulsoup4>=4.12
lxml>=4.9
apscheduler>=3.10,<4
jinja2>=3.1
prometheus_client>=0.19
